Fix build errors and harden .gitignore
Resolve three build failures under the modern CMake / VS 2026 toolchain:

- Set CMAKE_POLICY_VERSION_MINIMUM 3.5 so yaml-cpp 0.8.0 and glm 1.0.1
  (which require CMake < 3.5) configure under newer CMake.
- Bump CMAKE_CXX_STANDARD to 20; bgfx's bx library now uses designated
  initializers in its SIMD headers, which require C++20.
- Construct LayerConfig inside the try block in main.cpp instead of
  default-constructing it (its default ctor is intentionally private;
  instances must come from the validating factories).

Also clean up warnings and improve reproducibility:

- Rename Engine::update parameter deltaTime -> dt to stop it shadowing
  the Engine::deltaTime member (C4458).
- Add explicit uint16_t casts at bgfx::setViewRect call sites (C4244).
- Pin bgfx.cmake to release tag v1.143.9257-544 instead of master.
- Rewrite .gitignore to cover in-source CMake/MSBuild artifacts and
  FetchContent _deps; stop ignoring the tests/ source directory.

+++ src/renderer/BgfxRenderer.cpp
@@ -58,7 +58,7 @@ void BgfxRenderer::initialize() {
     }
 
     bgfx::setViewClear(0, BGFX_CLEAR_COLOR | BGFX_CLEAR_DEPTH, 0x303030ff, 1.0f, 0);
-    bgfx::setViewRect(0, 0, 0, viewWidth, viewHeight);
+    bgfx::setViewRect(0, 0, 0, static_cast<uint16_t>(viewWidth), static_cast<uint16_t>(viewHeight));
 
     VoxelVertex::initLayout();
     vbo = bgfx::createVertexBuffer(bgfx::makeRef(cubeVertices, sizeof(cubeVertices)), VoxelVertex::layout);
@@ -98,7 +98,7 @@ void BgfxRenderer::setViewport(int width, int height) {
     viewWidth  = static_cast<uint32_t>(width);
     viewHeight = static_cast<uint32_t>(height);
     bgfx::reset(viewWidth, viewHeight, BGFX_RESET_VSYNC);
-    bgfx::setViewRect(0, 0, 0, viewWidth, viewHeight);
+    bgfx::setViewRect(0, 0, 0, static_cast<uint16_t>(viewWidth), static_cast<uint16_t>(viewHeight));
 }
 
 void BgfxRenderer::setCameraPosition(const WorldCoord& pos) {
